\name{dataSupp-class}
\Rdversion{1.1}
\docType{class}
\alias{dataSupp-class}

\title{Class \code{"dataSupp"}}
\description{
La classe dataSupp
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("dataSupp", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{fonction}:}{Object of class \code{"function"} ~~ }
    \item{\code{decal}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "dataSupp" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("dataSupp")
}
\keyword{classes}
